moved sortedIdeaPages creation to idea-helpers file
