+++ src/commands/utility/idea.ts
@@ -15,37 +15,11 @@ import {
 import { ButtonComponent, Discord, ModalComponent } from "discordx";
 import { ARROW_BACKWARD_ID, ARROW_FORWARD_ID, BD4_BOT_ID, IDEA_TYPES, LIGHTBULB } from "../../constants.js";
 import { CommandConfig, Command, CommandCategory } from "../../types/command.js";
-import { getAllIdeas, UserIdea, UserIdeaTypedModel } from "../../types/data-access/idea.js";
+import { UserIdeaTypedModel } from "../../types/data-access/idea.js";
 import { ParentCommand, ParentCommandConfig } from "../../types/parent-command.js";
+import { refreshIdeas } from "../../util/idea-helpers.js";
 
-const ideas: UserIdea[] = await getAllIdeas();
-let ideaEmbeds: EmbedBuilder[] = [];
-const sortedIdeaPages: EmbedBuilder[][] = [];
-
-for (let x = 0; x < IDEA_TYPES.length + 1; x += 1) {
-	let pageNum = 0;
-	const shownIdeas: UserIdea[] = [];
-
-	for (let y = 0; y < ideas.length; y += 1) {
-		if (x === 0 || ideas[y].type === IDEA_TYPES[x - 1]) shownIdeas.push(ideas[y]);
-	}
-
-	for (let z = 0; z < shownIdeas.length; z += 1) {
-		// if i is perfectly divisible by 5 (ie 0, 5, 10), create new page
-		if (z % 5 === 0) {
-			if (z !== 0) pageNum += 1;
-			ideaEmbeds[pageNum] = new EmbedBuilder()
-				.setTitle(`${IDEA_TYPES[x] ? `${IDEA_TYPES[x]} Ideas` : "All Ideas"} | Page ${pageNum + 1}`);
-		}
-		ideaEmbeds[pageNum].addFields(
-			{ name: `(${z + 1})`, value: `Type: ${shownIdeas[z].type}\n Idea: ${shownIdeas[z].description}` }
-		);
-	}
-
-	sortedIdeaPages.push(ideaEmbeds);
-	ideaEmbeds = [];
-}
-
+let sortedIdeaPages = await refreshIdeas();
 
 const ideaSubmitConfig: CommandConfig = {
 	name: "submit",
@@ -149,6 +123,7 @@ const ideaListConfig: CommandConfig = {
 
 class IdeaListCommand extends Command {
 	public async run(msg: Message, args: string[]): Promise<boolean> {
+		sortedIdeaPages = await refreshIdeas();
 		enum Types {
 			all = 0,
 			utility = 1,
